[ci] support pypi
