+++ .github/workflows/publish.yml
@@ -73,6 +73,8 @@ jobs:
           pip install twine
       - name: package
         run: |
+          rm setup.py
+          ln -s setup_pypi.py setup.py
           python setup.py sdist
           twine check dist/*
       - name: Publish package pypi
